+++ e2e/consulcompat/shared_download_test.go
@@ -75,19 +75,6 @@ type consulJSON struct {
 	} `json:"versions"`
 }
 
-func usable(v, minimum *version.Version) bool {
-	switch {
-	case v.Prerelease() != "":
-		return false
-	case v.Metadata() != "":
-		return false
-	case v.LessThan(minimum):
-		return false
-	default:
-		return true
-	}
-}
-
 func keep(b build) bool {
 	exactVersion := os.Getenv(exactConsulVersionEnv)
 	if exactVersion != "" {
